reorganize code adn add back cpacity to add galaxies in padding in generate_dataset
